Refactor staff client components by extracting dialog components
Extracts inline dialogs from staff-works-client and staff-checkouts-client
into dedicated dialog components. Reduces parent components from 538/498
lines to 159/184 lines. Each dialog owns its own state and API logic.
