+++ src/app/staff/staff-checkouts-client.tsx
@@ -2,11 +2,8 @@
 
 import { useState, useMemo } from "react";
 import { useRouter } from "next/navigation";
-import { Check } from "lucide-react";
-import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Badge } from "@/components/ui/badge";
 import {
   Table,
@@ -17,28 +14,14 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from "@/components/ui/command";
+import { CheckoutFormDialog } from "./checkout-form-dialog";
+import { ReturnCheckoutDialog } from "./return-checkout-dialog";
 
 interface Checkout {
   id: string;
@@ -83,19 +66,6 @@ export function StaffCheckoutsClient({
   const [returningCheckout, setReturningCheckout] = useState<Checkout | null>(
     null,
   );
-  const [formError, setFormError] = useState("");
-  const [submitting, setSubmitting] = useState(false);
-
-  const [workListOpen, setWorkListOpen] = useState(false);
-  const [userListOpen, setUserListOpen] = useState(false);
-  const [workSearch, setWorkSearch] = useState("");
-  const [userSearch, setUserSearch] = useState("");
-
-  const [formData, setFormData] = useState({
-    work_id: "",
-    user_id: "",
-    due_date: "",
-  });
 
   const checkedOutWorkIds = useMemo(() => {
     const ids = new Set<string>();
@@ -125,22 +95,6 @@ export function StaffCheckoutsClient({
     });
   }, [checkouts, search, statusFilter]);
 
-  function defaultDueDate() {
-    const date = new Date();
-    date.setDate(date.getDate() + 30);
-    return date.toISOString().split("T")[0];
-  }
-
-  function openCheckoutDialog() {
-    setFormData({ work_id: "", user_id: "", due_date: defaultDueDate() });
-    setFormError("");
-    setWorkListOpen(false);
-    setUserListOpen(false);
-    setWorkSearch("");
-    setUserSearch("");
-    setFormOpen(true);
-  }
-
   function openReturnDialog(checkout: Checkout) {
     setReturningCheckout(checkout);
     setReturnOpen(true);
@@ -150,59 +104,10 @@ export function StaffCheckoutsClient({
     return !checkout.returned_at && new Date(checkout.due_date) < new Date();
   }
 
-  async function handleCheckout(e: React.SyntheticEvent) {
-    e.preventDefault();
-    setSubmitting(true);
-    setFormError("");
-
-    try {
-      const res = await fetch("/api/staff/checkouts", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(formData),
-      });
-
-      if (!res.ok) {
-        const data = await res.json();
-        setFormError(data.error || "Failed to create checkout");
-        return;
-      }
-
-      setFormOpen(false);
-      router.refresh();
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
-  async function handleReturn() {
-    if (!returningCheckout) return;
-    setSubmitting(true);
-
-    try {
-      const res = await fetch(`/api/staff/checkouts/${returningCheckout.id}`, {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ action: "return" }),
-      });
-
-      if (!res.ok) {
-        const data = await res.json();
-        setFormError(data.error || "Failed to return item");
-        return;
-      }
-
-      setReturnOpen(false);
-      router.refresh();
-    } finally {
-      setSubmitting(false);
-    }
-  }
-
   return (
     <div className="space-y-4">
       <div className="flex items-center gap-4">
-        <Button onClick={openCheckoutDialog}>Check Out Item</Button>
+        <Button onClick={() => setFormOpen(true)}>Check Out Item</Button>
         <Input
           id="checkout-search"
           placeholder="Search by title, name, or email..."
@@ -317,182 +222,20 @@ export function StaffCheckoutsClient({
         </TableBody>
       </Table>
 
-      {/* Check Out Dialog */}
-      <Dialog open={formOpen} onOpenChange={setFormOpen}>
-        <DialogContent
-          className="max-w-lg"
-          onOpenAutoFocus={(e) => {
-            e.preventDefault();
-            (e.currentTarget as HTMLElement).focus();
-          }}
-        >
-          <DialogHeader>
-            <DialogTitle>Check Out Item</DialogTitle>
-            <DialogDescription>
-              Select a work and a user to create a checkout.
-            </DialogDescription>
-          </DialogHeader>
-          <form onSubmit={handleCheckout} className="space-y-4">
-            <div className="space-y-2">
-              <span className="text-sm font-medium">Work</span>
-              <Command className="rounded-md border" shouldFilter={workListOpen}>
-                <CommandInput
-                  name="work-search"
-                  placeholder="Search works..."
-                  hideIcon={!workListOpen && !!formData.work_id}
-                  value={workListOpen ? workSearch : (availableWorks.find((w) => w.id === formData.work_id)?.title ?? "")}
-                  onValueChange={(v) => setWorkSearch(v)}
-                  onFocus={() => {
-                    setWorkSearch("");
-                    setWorkListOpen(true);
-                  }}
-                  onBlur={() => setTimeout(() => setWorkListOpen(false), 150)}
-                />
-                {workListOpen && (
-                  <CommandList>
-                    <CommandEmpty>No works found.</CommandEmpty>
-                    <CommandGroup>
-                      {availableWorks.map((work) => (
-                        <CommandItem
-                          key={work.id}
-                          value={work.title}
-                          onSelect={() => {
-                            setFormData((prev) => ({
-                              ...prev,
-                              work_id: work.id,
-                            }));
-                            setWorkSearch("");
-                            setWorkListOpen(false);
-                          }}
-                        >
-                          <Check
-                            className={cn(
-                              "mr-2 h-4 w-4",
-                              formData.work_id === work.id
-                                ? "opacity-100"
-                                : "opacity-0",
-                            )}
-                          />
-                          {work.title}
-                        </CommandItem>
-                      ))}
-                    </CommandGroup>
-                  </CommandList>
-                )}
-              </Command>
-            </div>
-            <div className="space-y-2">
-              <span className="text-sm font-medium">Borrower</span>
-              <Command className="rounded-md border" shouldFilter={userListOpen}>
-                <CommandInput
-                  name="borrower-search"
-                  placeholder="Search by name or email..."
-                  hideIcon={!userListOpen && !!formData.user_id}
-                  value={userListOpen ? userSearch : (() => {
-                    const u = users.find((u) => u.id === formData.user_id);
-                    return u ? `${u.name} (${u.email})` : "";
-                  })()}
-                  onValueChange={(v) => setUserSearch(v)}
-                  onFocus={() => {
-                    setUserSearch("");
-                    setUserListOpen(true);
-                  }}
-                  onBlur={() => setTimeout(() => setUserListOpen(false), 150)}
-                />
-                {userListOpen && (
-                  <CommandList>
-                    <CommandEmpty>No users found.</CommandEmpty>
-                    <CommandGroup>
-                      {users.map((user) => (
-                        <CommandItem
-                          key={user.id}
-                          value={`${user.name} ${user.email}`}
-                          onSelect={() => {
-                            setFormData((prev) => ({
-                              ...prev,
-                              user_id: user.id,
-                            }));
-                            setUserSearch("");
-                            setUserListOpen(false);
-                          }}
-                        >
-                          <Check
-                            className={cn(
-                              "mr-2 h-4 w-4",
-                              formData.user_id === user.id
-                                ? "opacity-100"
-                                : "opacity-0",
-                            )}
-                          />
-                          {user.name} ({user.email})
-                        </CommandItem>
-                      ))}
-                    </CommandGroup>
-                  </CommandList>
-                )}
-              </Command>
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="due_date">
-                Due Date{" "}
-                <span className="text-muted-foreground font-normal">(30 days by default)</span>
-              </Label>
-              <Input
-                id="due_date"
-                type="date"
-                required
-                value={formData.due_date}
-                onChange={(e) =>
-                  setFormData((prev) => ({
-                    ...prev,
-                    due_date: e.target.value,
-                  }))
-                }
-              />
-            </div>
-            {formError && (
-              <p className="text-sm text-destructive">{formError}</p>
-            )}
-            <DialogFooter>
-              <Button
-                type="button"
-                variant="outline"
-                onClick={() => setFormOpen(false)}
-              >
-                Cancel
-              </Button>
-              <Button
-                type="submit"
-                disabled={submitting || !formData.work_id || !formData.user_id}
-              >
-                {submitting ? "Checking out..." : "Check Out"}
-              </Button>
-            </DialogFooter>
-          </form>
-        </DialogContent>
-      </Dialog>
+      <CheckoutFormDialog
+        open={formOpen}
+        onOpenChange={setFormOpen}
+        availableWorks={availableWorks}
+        users={users}
+        onCreated={() => router.refresh()}
+      />
 
-      {/* Return Confirmation Dialog */}
-      <Dialog open={returnOpen} onOpenChange={setReturnOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Return Item</DialogTitle>
-            <DialogDescription>
-              Return &quot;{returningCheckout?.work_title}&quot; from{" "}
-              {returningCheckout?.user_name}?
-            </DialogDescription>
-          </DialogHeader>
-          {formError && <p className="text-sm text-destructive">{formError}</p>}
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setReturnOpen(false)}>
-              Cancel
-            </Button>
-            <Button onClick={handleReturn} disabled={submitting}>
-              {submitting ? "Returning..." : "Confirm Return"}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <ReturnCheckoutDialog
+        open={returnOpen}
+        onOpenChange={setReturnOpen}
+        checkout={returningCheckout}
+        onReturned={() => router.refresh()}
+      />
     </div>
   );
 }
